Implement a very basic OIDC auth flow
This has some problems:

1. It goes through the OIDC flow for every request to /manage
2. It doesn't have a separate callback page which redirects to
   /manage, so the URL gets a bunch of ugly parameters
3. Because of those URL parameters refreshing the /manage page doesn't
   work!

I think ideally we'd use the "code" flow, and have a separate callback
controller which uses the auth code to fetch the ID and auth tokens,
and stores them in a session.  Then we only need to re-auth when the
session expires.

I think, however, that for account management we might *not* want to
use OIDC for fetching the ID info!  The use-case for this app is a bit
different to the normal.

Normally, an app which needs identity info uses the OIDC auth flow to
request what it needs in the ID token, and then it uses that ID token
to do the rest of its work.  But for account management, we need to be
able to change attributes (which you can't do through OIDC) and always
see the latest state (which would require fetching a new ID token on
every request).  We need to always see the latest state because
otherwise having the ability to change state is too risky.

So I think for the management app, we want to log in with OIDC, but
then use the Keycloak API to interact with the user's data directly.

For, eg, the use-case of auto-selecting options in a checker OIDC and
a single ID token will work fine.