Use Keycloak API to fetch user info & add callback route
Following on from the argument in the previous commit, this one
ditches the use of the ID token (though I've left it in the code as an
example of how to get it from the OIDC response), but only store the
subject ID in the session, and then use the Keycloak API directly.

We'll probably want to implement a timeout for the session state.