Move Dotenv to Development and test
This is a bit of a conflict, turns out both I and another dev had added
this. I'll leave my case for adding it below even though it's already
been agreed to. Moved it to development and test as it might still be
useful if we wanted to run it without docker-compose. Also generated
binstubs

As we will have to manage envars for keycloak, dotenv will make it much
easier to write integration tests

Add dotenv binstub