Implement POST method for requesting password reset email
This adds functionality onto the password reset form, so a password
reset email is sent out (using GOV.UK Notify) when the form is
submitted.

When the form is submitted, we produce a random token which is written
to the user's attributes.  This token will be used later to determine
whether a request to reset a password is authentic.