POST part of allowing users to set a new password
This actually makes the change of password to a user's account, as well
as implementing some basic error handling.

We need to check the authenticity of the token again (by means of
putting it in hidden form fields in the reset form) to ensure the POST
request is made by the genuine user (rather than by someone who changes
their `user_id` in the request parameters).