Fix configuration errors in concourse deploy
Fixes the following:
- My linter had said that all git source URIs in resources should end in
.git. I'd changed that but left it pointing to the HTTPS source. URI's
for git resources should use the proper git@github ssh route

- Renamed and uploaded a correct private key to gain access to the
private repo (had to do it with hijack there may be a bug that adds
spaces instead of newlines in gds-api - will raise an issue).

- Having refactored the app instance out of the task and into a pipeline
variable, I needed to prepend it with a $ to be correctly interpolated
into the command.

- Corrected various naming errors

Amend concourse git link to use ssh protocol

Also a helpful test to see that when we merge into the main branch the pipeline updates itself