Fix HTML injection issues
`sanitize` makes a string HTML-safe by stripping harmful tags.  This
is ok when it's sensible HTML in our locale files, but not so great
when it's user-controlled input like an email address.  We don't want
users to be able to insert `<p>` or `<strong>` tags into the page.

So, remove most of the calls to `sanitize` and use `h` on user-input
which is interpolated into the few remaining `sanitize` calls.