Finish change-of-email implementation
We probably want to discuss how this should work.  This is what I've
gone for so far:

1. User submits the form
2. Email with a cancel link sent to the current address
3. Email with a confirmation link sent to the new address

Then, if the user clicks the cancel link *before* the confirm link the
change is cancelled.  If they click it *after* the confirm link,
they're too late.

The email address isn't actually changed until the confirm link is
clicked, so we have to check it's available at that point.  We could
check up front as well, but the check at point of confirmation will
still be needed: unless initiating a change "reserves" the email
address, but that seems open for abuse.

---

Logging in to keycloak is done with either the username or the email
address, the OIDC sub is a UUID distinct from either, and in this app
we look up users by email address.  So I don't think we need to worry
about updating the username just yet.