Add password reset feature
And confirm to the user when one has been sent.

I thought for a while about what to do if the User cannot be found in the database, I think in _any_ case we should return the same screen and possibly tweak the content to be a little more non-committal about if something sent or not. Otherwise you would be able to use this form to query if an email address has a GOV.UK account which would be a trade off between a small user convenience for a more serious data leakage. Will bring this up at next show and tell