Put OIDC secrets in secrets.yml
Rather than read from ENV in the initialiser, all access to secrets
should go through one place.

I've also removed the support for reading a private key file in
production, that now has to come from an environment variable, and
renamed the "salt" to "pepper", as it's not random per-hash.