Change from allowing scopes to denying them
We're likely to have many more scopes that we want to allow than deny (possibly just :openid that should be denied !). So it makes sense to switch this from an allow list to a deny list

This work has cropped up as I'm experimenting with adding some scopes along the lines of :'emailsubscriptions.read_only' and :'emailsubscriptions.write_only'. I'm alreadying having to add that in my factories, and in /config/initalizers/doorkeeper.rb. It would be nice to not have another place to maintain these!