Configure peppering for devise
Passwords can be made more secure by adding salt and pepper values
before hashing:

- Salt: a random value, stored in the database with the hash
- Pepper: a constant value, stored as a secret

Both salting and peppering protect against rainbow tables (lists of
precomputed hashes).

Imagine your hashes are unsalted and unpeppered, and a database leak
occurs: an attacker can simply find a rainbow table for whatever
hashing algorithm you're using, and look up hash values to recover
passwords.

So to solve that problem you add a random salt to each hash.  You need
to store the salt in plaintext, so you can verify hashes, but that's
fine.  An attacker now knows "I need a rainbow table for <hash> where
every password has the salt value prepended/appended/etc", but the
salt value is different for every password, so the attacker needs a
different table for each salt.  This is much more secure.

Eventually though, these salted rainbow tables start to proliferate.
So to mitigate against those, we add a new secret value to the hash.
This is the pepper.  The attacker may know the hash function, and the
salt, and the password hash, but without knowing the secret pepper
value they cannot tell which rainbow table to use, and have to
brute-force the hash.

Peppering is an addition to salting, not a replacement.  If salting
were not used, then password hashes would be just:

    hash(password + pepper)

So if an attacker has long-term access to the database, they could
experiment by setting a password to multiple known values, inspecting
the hashes, and trying to reverse-engineer the pepper.  Salting makes
this attack too expensive, by adding a random component which changes
whenever the password does.