Load scopes into Doorkeeper from config file
We're likely going to end up with a lot of scopes, and maintaining them in multiple places is likeing to be boring. On top of that we may actually want to share a config between the attribute store and this file. So as a first step, we've taken these out of the doorkeeper initalizer and put them into a yaml file in config which then get's loaded in by a small lib class.

Perhaps later we could load this in as a gem