Use application uid in pairwise subject identifier generation
Currently we're using

    hash(user + application uri + pepper)

Which means that an application can't change its uri, as all the user
identifiers will change.  Instead, use

    hash(user + application id + pepper)

The application ID won't change, so this is a more sensible hash.

The OIDC spec doesnt define how a pairwise subject identifier should
be computed, it only provides examples, so changing our algorithm is
fine.