Disable basic auth for deanonymise_tokens endpoint
We have no basic auth on any of the OAuth / OIDC endpoints, as
Doorkeeper's controllers don't inherit from ApplicationController by
default.  This one does, so it's an inconsistency.

Rather than make the Attribute Service able to handle HTTP Basic Auth
on this one API route, just make it consistent with the others.