Add a few tests around user tokens being expired or revoked
We don't currently expire or revoke the tokens used for communicating
with the Attribute Service, but I imagine we might like to one day:

1. If there's a bug which leaks access tokens, we'd want to revoke them
   all to issue new tokens.

2. To mitigate 1, we might want to set an expiry time on tokens.

So I thought there should be some tests making sure we correctly issue
new tokens when old ones are no longer valid.  This revealed a bug: if
only expired tokens exist, we wouldn't create a new one.