Redirect a logged in user from / to /account
Previously, visiting / while logged in would give you the email
address form.  Entering your email address would take you to the login
form (because the account exists), which would then redirect you to
the accounts page (because you have a valid session).

This is a bit confusing, as it looks like you're logging in without
entering a password.

So this commit pushes the redirect to the / page directly.