Add page asking if users want email notifications
This works by implementing a small state machine in the devise
registrations controller, deciding which page to show when the #create
endpoint is hit.

This is a bit complicated, but I'm not sure what a better solution
would be: we could create the account but in some sort of
"deactivated" state, and store the JWT alongside it, and then make it
so that trying to use a deactivated account sends you to the T&C page
and then the email alert consent form, but that introduces problems:

- can we even store the user's email address and attributes before
  they see the T&Cs?
- the login needs additional logic
- we'd have to stop devise from sending its confirmation email until
  the user has finished the registration process

So this complicated state machine may be a lesser evil for now.