Make registration stateful
We're going to need to bring the database into our registration
process as soon as we want to do SMS 2fa, as putting the desired code
in a hidden form field is a bad idea.

This commit removes the RegistrationStateMachine, splitting it out
into separate controllers and views.  State is persisted with a new
RegistrationState model, and a registration_state_id parameter is
passed around as a reference to this.

I started by using the session to hold the state ID, but that makes
more awkward.