Track which form a user should see in the state
Previously, a user could bypass parts of the multi-page form by
editing the form target.  This isn't really a problem here, but when
we start to implement SMS 2fa it will be - we don't want a user to be
able to skip that by just setting the form target to the finish page.

So add a field to the RegistrationState which tracks the page a user
should be on, and redirect them if they end up on the wrong page.

I had to override 'respond_with', which Devise uses to do conditional
redirects, as its location: parameter only applies for POST requests,
but now the page is reached via a GET request (and making it a POST
request messes up all the pages which can reach it).