Revise email sigup routes
We're replicing the singular devise registration edit route (that used to do
both email and password) with two routes that deals with each separately.

Registrations are an out-of-the-box controller that comes from devise:
https://github.com/heartcombo/devise/blob/master/app/controllers/devise/registrations_controller.rb

Where they bundle it into one form, we're going to make two, which we can do by
overriding their defaults in a controller of our own.

The old edit path will continue to exist as it's a default in their controllers
but having removed the route it won't be accessible from our app