Add devise prepend_before_action for new methods
The old devise edit route was listed in a prepend_before_action at the
top of the file. This allowed it access to the `resource` object which
we'll need here too.

See here:
https://github.com/heartcombo/devise/blob/master/app/controllers/devise/registrations_controller.rb

My assumption is that anything declared here will overwrite the entire
line from the default controller, so I want to make sure we've included
those other routes that also need scope authenticated and minimum
password length checked (new/destroy).

However this makes rubocop very sad as it. Fairly reasonably, it expects
you only to work with methods that are defined here in the class, and
it's not picking up on the inheritance.

So I've stopped it complaining with a comment as I'm pretty sure this is
the way to go.