Add form validation for Feedback from
Requires changing the existing error helper which was taking data out of
the flash hash, instead we can get this from an instance variable handed
over from the controller.

The frontend components require an href key to exist in the data
structure, so that has to stay. It didn't make much sense to me to run
the select off of that, so whilt there's a bit of duplication there
having both href and field, I figured that was cleaner in to keep in the
error helper