Fix bad references to translations
These entries are under `general`, not `account.general`.  The reason
this didn't show up in the views is because if the translation doesn't
exist, Rails defaults to the capitalised form of the last component of
the name, which was coincidentally correct.

Also `yes` and `no` are interpreted as booleans in YAML unless
quoted (even in key names).