Update registration_state.phone and send code in a transaction
We don't verify that the phone number is the same when validating a
code, so we could potentially end up in the state where:

1. A user enters a phone number they control and gets sent a code
2. A user enters a phone number they don't control:
  1. The registration_state is updated to store the new number
  2. An error prevents the new code from being generated
3. The user enters the old code, claiming ownership of the new number

So wrap the updating of the registration_state and the generating of
the code in a transaction.  If the code isn't generated due to an
error, the number won't be updated either.