Import Devise Registration destroy controller
Where we want to change the default behaviour of devise we can bring
it's controllers across and it'll use ours instead of the one from the
gem.

In this case it's identical except we've changed the path we route the
user to after delete (their default would be root, we want to go
elsewhere).

Also had to make a few slight modifications to appease rubocop, it
wanted the destroy! command for anything where the return value wasn't
checked and a space between some parents and curly brackets.

I took the idea of using the safe navigator operator from the
registrations destroy below.