Record DeanonymiseToken uses as "data activities"
Currently we use access grants for showing when data has been shared,
but an access grant only shows when a token was granted, not when it
was used.  Tokens can be used long after they were created, and
re-used if they get refreshed.  So it's better to show token use than
token creation.