Tell users when their email address has been confirmed
When a user confirms their email address, we currently don't tell
them.  They have to look at their attributes and notice the
"(unconfirmed)" is gone.  Even worse, because we don't log a user in
when they confirm their address (which is desirable to avoid a
loophole where 2fa can be avoided), a user might end up on the login
screen with no indication that anything has happened, click the link
again, and be told that they've already confirmed their account!

This looks pretty bad on the login form currently, but that will be
resolved by adding a proper heading to the page, which we need to do
anyway.