Deduplicate similar data or security activities
Two data activities are similar if they're from the same application
and occurred within a minute of each other.

Two security activities are similar if all the attributes are the same
and they occurred within a minute of each other.

The intent of this is to reduce noise from (eg) applications using an
OAuth token multiple times (which currently show up as multiple data
activities) or zealously redirecting a user through the OAuth login
flow multiple times.