Pass cookie consent back to the transition checker
This is a hacky solution which works for now but which we will want to
revisit soon: if the user has consented to cookies *and* the
previous_url takes them back to the transition checker, append
something to the state.

This is not good because it couples us to the representation of the
transition checker state, and will also break if the query params get
re-ordered.

A better solution would be storing cookie consent as an attribute, but
currently we only want to implement per-session consent, not
per-account consent, so that doesn't work.