Update content for MFA phone number create
Requires a new bottom link that allows the user to bail out and get back
to the email alerts only path. As this is a complex URL and needs to be
plucked out of the JWT I've put this in a helper.

Also added a new class to the form and put a margin-bottom on.