Sign out of all sessions
Devise uses a salted cookie to keep a user logged in.  By default,
this salt is just the hashed password, so the only way to sign out of
other sessions is to change your password.

This commit adds a random value to the salt, which is refreshed on log
out, to prevent session reuse.  This change doesn't prevent concurrent
sessions, and if a session times out others will be unaffected; but if
the user logs out of one session, all the others will be terminated
too.

We could extend this to prevent multiple sessions by also refreshing
the salt on login.