Keep user on correct form on email or password change error
We are using separate views for changing the email and password, whereas
devise has a single view for both.

By default, we are directing users to the combined form when an error
occurs rather than keeping them on the correct form. This overrides that
behaviour to ensure users remain on the same form, but with the error
message showing.