Set path=/ for cookies_policy
From RFC 6265 (HTTP State Management):

> 11. If the cookie store contains a cookie with the same name,
>     domain, and path as the newly created cookie:
>
>     1. Let old-cookie be the existing cookie with the same name,
>        domain, and path as the newly created cookie.  (Notice that
>        this algorithm maintains the invariant that there is at most
>        one such cookie.)
>
>     2. If the newly created cookie was received from a "non-HTTP"
>        API and the old-cookie's http-only-flag is set, abort these
>        steps and ignore the newly created cookie entirely.
>
>     3. Update the creation-time of the newly created cookie to
>        match the creation-time of the old-cookie.
>
>     4. Remove the old-cookie from the cookie store.
>
> 12. Insert the newly created cookie into the cookie store.

We're already setting a default cookie (with the cookie banner), and
currently this new cookie we're setting on sign-in isn't overwriting
the old one.  I suspect the issue is the path, as Firefox dev tools
shows that the default cookie has a path of "/", whereas we're not
setting a path at all for our new cookie.