Set consent cookie in registration in POST your-information response
Currently, we still display the cookie banner on the email alerts
page (for users who sign up through the transition checker), because
we're setting the cookie only on the finish page.

So to fix that, set the cookie in the POST response.

I renamed the parameter of the cookies_policy_header helper because it
doesn't have to be a user, just something with a cookie_consent
method.