Log out of the checker on delete
Redirect to checker?continue=delete then, in the logout controller,
check if ?done == delete and if so redirect to the post-delete
confirmation page.

How deleting your account works with services which may have a long
session time is something we'll need to figure out.  The service will
notice as soon as it tries to fetch any of your attributes, but if it
just uses the session cookie to persist some service-specific state it
may be a while before the user is logged out.