Pluck email out of user model on post-registration page
The user exists at this point, and gets logged in at the end of
registration, so there's no need to pass the email address along in
the URL.  This also removes a place where we're logging PII.

And as an additional bonus, we get to remove some copy&pasted Devise
code, which we currently have too much of.

To ensure current_user is not nil (to prevent people from going
straight to the page), check the user is authenticated in the
controller.