Present phone number to Notify in E.164 standard format
This will ensure all user input formatting is removed and the number is
presented in a standard format.

Note that this only accepts UK numbers, so we will need to remove this
line when international support is added (in that case, we should be
storing the E.164 number in the database instead).