Use standardised phone number when updating phone number
In addition to ensuring the number is stored consistently, this gives us
another advantage: the number is compared regardless of the user's
formatting. E.g. if the user entered a space first time, but then tries
to change to the same number without a space, we would have previously
considered this to be a different phone number - this will no longer be
the case.