Only report account manager logins as a "login" in stats
A "login" event covers logging into the account manager and also
logging into a service, but it's only in the former case when the user
will actually enter some credentials.  This is confusing to
non-devs (which is who these statistics are intended for), so only
count "a user enters some credentials" as a "login" here.