Support international phone numbers without a plus prefix
Some users enter their international dialing code with 00 instead of a +
(e.g. 00353 instead of +353 for Republic of Ireland). This is because
that is how the number is dialled on a landline telephone.

This format was not considered valid by the telephone_number gem, so we
need to manually substitute the 00 prefix with a + in these cases.