Filter sensitive POST and GET params from logs.
Last time we tried this, it didn't go quite to plan[1][2]. So this time
around I'm going to try something a bit more gradual.

This will be phase 1. Which is to just keep our existing logs in the
same format as currently but redact out any potentally sensitive PII.

In step 2 we will start streaming these to Splunk for auditing.
Filtering is so that they only see the information required.

As a final step 3, I will return to trying to format the logs as JSON,
this should make it easier for the cyber folks to query what is going
on. This, via a gem was the step that seemed to be causing issues last
time around.

[1]: https://github.com/alphagov/govuk-account-manager-prototype/pull/430
[2]: https://github.com/alphagov/govuk-account-manager-prototype/pull/437