Add tests for Crown Dependency mobile numbers
These are formatted as UK numbers, with a +44 area code, but are
considered international (and therefore don't validate as UK numbers).
Adding a test to make sure they are supported.

For reference, the area codes used are:
- Guernsey: 07781, 07839 and 07911
- Isle of Man: 07624
- Jersey: 07797, 07700 and 07829

Note that at the time of writing, 07797 has only been issued as 07797 8
and 07797 9, so the test cannot use 123456 for this number. Equally
07624 has not been issed to 07624 8 or 07624 9 yet, so we cannot used
the same six-digit number for all area codes.