Treat Crown Dependency phone numbers as UK numbers
The Crown Dependencies (Guernsey, Isle of Man and Jersey) use the UK
country code (+44) but have their own area codes and are treated as
international numbers. Because of this, they currently fail our
validation and are not accepted.

This change will allow us to treat Crown Dependency mobile numbers as if
they are UK numbers.