Deduplicate logins when dumping statistics
When we show users the security activities on their account, we are
de-duplicating logins that occur within 60 seconds on each other. We
were not doing that for the daily statistics.

Removing these will give us greater confidence in the statistics since
we know that certain things (like being asked for a MFA code) create
another login event.