Push confirmation parameters into session
The email confirmation sent page is reached from three places:

- At the end of the registration journey
- After changing your email address
- From resending a confirmation email

The benefit of having a single controller / view is that the pages
don't get out of sync.  But it does mean that this page needs to know
whether it should display the email address of the current user, or
something else.

Previously this was done through query params.  Now it's done through
session variables.  This removes the email, user_is_confirmed, and
new_user query params.