Prevent sending two checker confirmation emails
Depends on: https://github.com/alphagov/gds-api-adapters/pull/1011

We are going to reinstate sending the user a confirmation email,
which contains a link to their Brexit Checker results. However, this
would conflict with the email sent by accounts, and it was only a
coincidence that we removed it just before accounts launched [1].
Since the email sent by this app is more relevant for a subscription
created via an account, this modifies the "subscribe" API call to
instruct the email system not to send its own confirmation email.

[1]: https://github.com/alphagov/email-alert-api/pull/1465