Move JWT outside of registration and login state models
This will allow us to use a JWT for either a login state or registration
state, or both. Therefore we can pass a user around the login and
registration forms by simply setting the JWT as a cookie.

Use of `delegate` in the models means we don't need to update much of
the existing code that use the JWT.

The migration will ensure users currently in registration flows are able
to continue unimpeded. We use `read_attribute` in this migration to
bypass the newly introduced `delegate` in the RegistrationState model
and read this attribute directly from the model, rather than delegating
to Jwt.

We will expire the Jwt after 60 minutes only if it is not attached to a
login or registration state. If it is attached to either, it will be
expired only when the attached state is also expired. This means the Jwt
will persist for longer than 60 minutes if the user is part way through
a session.