Log security events to stdout
This will send them to Splunk, which is monitoring our application
logs.

The "notes" field is omitted as it contains PII (so is the user's
email address, for the same reason).  If we notice suspicious activity
from Splunk analysis, we can look up the relevant notes in the app
database.