Skip parsing JWT payload on migration
We parse the raw JWT on saving the Jwt model and stored the parsed data
in the database. However in the migration, we already have parsed data
so don't need to parse this again.

This wasn't caught previously as the migration had been run before the
before_save callback was introduced to the Jwt model.