Destroy stale login and registration states upon success
If a user has a JWT and they attempt to signup or login multiple times,
we will associate the Jwt object with multiple login or registration
states. This results in a foreign key failure when they successfully do
one of these actions and we attempt to delete the state.

By manually deleting these stale states, we stop this happening.