Return jwt from get_payload instead of string
It looks like the get_payload method is supposed to:

* parse the string input if provided
* lookup a pre-existing session if there's no input
* return a HashMap with symbolized keys representing the JWT

However, if you provided it with a string as input, it would return the
same string as the return value, instead of the parsed JWT.

There were no tests covering the (weird?) situation where someone who's
already signed in signs in again with a fresh JWT, so this wasn't caught
until production:

https://sentry.io/organizations/govuk/issues/2098313434/?project=5370953

I've added a little test that would have reproduced the issue.

I suspect that having the parameter called `payload` contributed to this
mistake, so I've renamed it to `payload_string` to make it a bit clearer
(although I don't normally advocate sticking types in variable names).