Add test for JWT user registering twice
In d043c9f, we fixed a problem where a user (coming from another
application with a JWT) would go part way through the registration then
start registering again without returning to the application. This
resulted a Jwt attached to multiple RegistrationStates. Due to the
dependency, destruction of the Jwt failed when the user successfully
registered. This adds a test to cover the scenario.