Simplify welcome controller / get_payload
This PR makes three changes to the JWT handling:

get_payload does two different things, and one of those things is only
used by a single controller, so just inline that branch where it's
used.

If there is a jwt_id in the session, but no Jwt model is found, delete
the session key.

JWTs are sent by cross-domain POST requests, so if a JWT is received
in the welcome controller, the user will be logged out, meaning the
logged-in branch can be simplified to not refer to the JWT.

This also solves what is, I think, a bug:

1. The user arrives on the welcome controller with a JWT
2. jwt_id gets set in their session
3. They log in, and get sent to the JWT's post-login path
4. Some time later, but before the session has expired, the user does
   a GET request to the welcome controller
5. The user gets sent to the JWT's post-login path again