Use 'match: :first' to find "your information" checkboxes
We've seen this test sometimes failing with the element not found,
which suggests the page is being slow to load.  I can't reproduce this
locally, only in GitHub Actions or in Concourse, so it's probably due
to load on those while this test is running.

Using `match: :first` will make Capybara wait up to 2 seconds before
erroring.