Correctly order RegistrationStates for passing 2fa in test
The registration from another app test was flaky.  Initially I thought
it was a timing issue with that page, but adding a delay didn't help.
I then noticed that this is the test which creates two
RegistrationStates, and the step failing is potentially the 2fa
validation step.  Could it be using the wrong code?

As it turns out, yes:

    > RegistrationState.order_by(:touched_at).pluck(:id)
    ["84471eb6-2e68-447e-a16b-d6d68624eb4c",
    "46085b18-558a-4744-a503-b9e3283dce09"]

    > RegistrationState.last.id
    84471eb6-2e68-447e-a16b-d6d68624eb4c

When not specifying an order, things are implicitly ordered by primary
key.  And you can't trust the ordering of randomly generated UUIDs.

I've also added the fix to the normal registration specs, to avoid
this same issue cropping up in the future.