Add rake task to report on commonly used passwords
We have recently implemented a block on new users using a password on
the NCSC's commonly used password list [1].

However we believe existing users may have already used these passwords.
This task will report how many users this affects, so we can plan on how
to ask them to change their password.

Since the task is slow and could be terminated by a deployment, it is
run as a Sidekiq job in a separate queue (with the lowest priority), so
as not to block any other queued tasks. If re-run, it will only check
users who do not have an existing check result stored.

1: https://github.com/alphagov/govuk-account-manager-prototype/pull/495