Do not throw an error for `is_password_banned?(nil)`
We've seen this error a few times in the registration controller.
It'll happen if the form is submitted without the field being sent.
Note, this is different to the field being sent with an *empty*
value.

So I think this will only happen if:

- the browser is doing something odd (not sending empty fields), or
- the user is constructing the request directly (eg with curl)

Since we've only had 24 errors in 5 days, it doesn't seem like the
problem is triggered by form submissions with an empty password.

While `nil` is not a *banned* password, it is still invalid because
it's forbidden by our length requirement.