Refactor existing security expect event method
The method assumed that the user would always be the factory at the top
of the test file. This works for security events affecting an existing
user.

However we want to add a test that checks for a security event on user
creation. So this adds the optional ability to pass a user explcitly to
be matched in the security events.

The number of optional parameters was getting quite long for the method
so I switched it for a splat'd options hash.

Defaults are then merged in from a new method below.