Look up the country of an IP when recording an event
The country is stored in the Rails cache for a day, to avoid repeated
lookups for future events.  I decided to use the cache, rather than
having a table mapping IP addresses to countries, because IP addresses
do move between countries on occasion, so we want to have somewhat
up-to-date data.