Fix inactive users not seeing any error message at log in
Or, do not call failure app on session create when user is inactive.

Rather than checking if a user is inactive before authenticating them,
Devise uses a Warden hook to boot them out *after* authenticating
them[1].  This is reasonable, because the alternative is that every
strategy needs to implement the check, and it's something that could
easily be overlooked.

When a user is booted out in this way, the failure app gets called,
which sets a flash alert and 302-redirects the user back to the new
session path.  This also makes sense, Devise uses a lot of flash
alerts, so it's a standard pattern.

However we've customised our sign in process a lot, and in particular
we're not using flash alerts for any of our errors.  So the standard
approach doesn't quite work for us.

So what happens (prior to this change) is:

- An account exists which hasn't been confirmed in time.
- The user tries to log in:
  - If they enter the correct password:
    - They get 302-redirected back to the log in form, and don't see
      any error messages.
  - If they enter an incorrect password:
    - A 200 response is rendered which displays the error message.

This is unhelpful, and inconsistent, and confusing.  We want to show
the error message in all cases.

Rather than try to untangle our custom behaviour and use flash alerts
more, I decided the easiest work-around is to implement the inactivity
check as part of the auth strategy.

[1] https://github.com/heartcombo/devise/blob/master/lib/devise/hooks/activatable.rb